Add Progress Bar.

+++ Src/Application/ReliefApp.h
@@ -24,7 +24,7 @@ namespace MagicApp
         ~ReliefApp();
 
         virtual bool Enter(void);
-        virtual bool Update(float timeElapsed);
+        virtual bool Update(double timeElapsed);
         virtual bool Exit(void);
         virtual bool MouseMoved(const OIS::MouseEvent &arg);
         virtual bool MousePressed(const OIS::MouseEvent &arg, OIS::MouseButtonID id);
